Revert refactor: Put `fetch_crate_cratesio` back to `drivers`
